Corrected inheritance class names and removed demo code
